Improve switch and settings tabs accessibility
We had a lot of a11y warnings on build. We still have some now, but not
that many any more. The settings tabs are now navigable by keyboard and
should be announced correctly. The solution is not perfect (yet), as
BeerCSS requires some things to be done that way, unfortunately.